implemented fallback player
