+++ web/src/index.html
@@ -20,7 +20,8 @@
 
         <div class="main">
             <div class="playback">
-                <video controls>
+                <!-- <video controls> -->
+                <video>
                     Your browser doesn't support HTML5 videos.
                 </video>
             </div>
